Refactor: 멘토님 코멘트 리팩토링
