+++ src/components/main/Sider.tsx
@@ -25,6 +25,29 @@ function Sider({ isOpen, handleClose }: MainSiderProps) {
 	const navigate = useNavigate();
 	const [isAccessToken, setIsAccessToken] = useState(checkAccessToken());
 	const [date, setDate] = useRecoilState(dateState);
+	const handleMyPageClick = () => {
+		if (isAccessToken === false) {
+			removeCookie();
+			Swal.fire({
+				title: '로그인이 필요한 서비스입니다.',
+				text: '로그인 창으로 이동하시겠습니까?',
+				icon: 'warning',
+				showCancelButton: true,
+				confirmButtonColor: '#3085d6',
+				cancelButtonColor: '#d33',
+				confirmButtonText: '확인',
+				cancelButtonText: '취소',
+			}).then((result) => {
+				if (result.isConfirmed) {
+					navigate('/login');
+				} else {
+					Swal.close();
+				}
+			});
+		} else {
+			navigate('/mypage');
+		}
+	};
 	return (
 		<Drawer
 			placement="left"
@@ -97,29 +120,8 @@ function Sider({ isOpen, handleClose }: MainSiderProps) {
 			<List>
 				<ListItem
 					onClick={() => {
-						const res = checkAccessToken();
 						handleClose();
-						if (res === false) {
-							removeCookie();
-							Swal.fire({
-								title: '로그인이 필요한 서비스입니다.',
-								text: '로그인 창으로 이동하시겠습니까?',
-								icon: 'warning',
-								showCancelButton: true,
-								confirmButtonColor: '#3085d6',
-								cancelButtonColor: '#d33',
-								confirmButtonText: '확인',
-								cancelButtonText: '취소',
-							}).then((result) => {
-								if (result.isConfirmed) {
-									navigate('/login');
-								} else {
-									Swal.close();
-								}
-							});
-						} else {
-							navigate('/mypage');
-						}
+						handleMyPageClick();
 					}}
 				>
 					<ListItemPrefix>
