+++ src/types/ReservationCard.type.ts
@@ -1,0 +1,11 @@
+export interface reservationCardProps {
+	code: string;
+	accomodationName: string;
+	accomodationImage: string;
+	checkinDate: string;
+	checkoutDate: string;
+	isUsed: boolean;
+	accomodationId: number;
+	capacity: number;
+	roomName: string;
+}
